+++ gui/windows/RunPage.cpp
@@ -144,7 +144,10 @@ wxWindow* RunPage::createParticleBox(wxPanel* parent) {
     const Float cutoff = gui.get<Float>(GuiSettingsId::CAMERA_ORTHO_CUTOFF) * 1.e-3_f;
 
     FloatTextCtrl* cutoffCtrl = new FloatTextCtrl(particleBox, cutoff, Interval(0, LARGE));
-    cutoffCtrl->onValueChanged = [this](const double value) { this->updateCutoff(value * 1.e3_f); };
+    cutoffCtrl->onValueChanged = [this](const double value) {
+        this->updateCutoff(value * 1.e3_f);
+        return true;
+    };
     cutoffCtrl->SetToolTip(
         "Specifies the cutoff distance in kilometers for rendering particles. When set to a positive number, "
         "only particles in a layer of specified thickness are rendered. Zero means all particles are "
@@ -165,6 +168,7 @@ wxWindow* RunPage::createParticleBox(wxPanel* parent) {
     particleSizeCtrl->onValueChanged = [this](const Float value) {
         gui.set(GuiSettingsId::PARTICLE_RADIUS, value);
         controller->tryRedraw();
+        return true;
     };
     particleSizeSizer->Add(particleSizeCtrl, 1, wxALIGN_CENTER_VERTICAL);
     particleSizeSizer->AddSpacer(boxPadding);
@@ -246,6 +250,7 @@ wxWindow* RunPage::createRaymarcherBox(wxPanel* parent) {
         GuiSettings& gui = controller->getParams();
         gui.set(GuiSettingsId::SURFACE_LEVEL, value);
         controller->tryRedraw();
+        return true;
     };
     levelSizer->Add(levelCtrl, 1, wxALIGN_CENTER_VERTICAL);
     levelSizer->AddSpacer(boxPadding);
@@ -261,6 +266,7 @@ wxWindow* RunPage::createRaymarcherBox(wxPanel* parent) {
         GuiSettings& gui = controller->getParams();
         gui.set(GuiSettingsId::SURFACE_SUN_INTENSITY, value);
         controller->tryRedraw();
+        return true;
     };
     sunlightSizer->Add(sunlightCtrl, 1, wxALIGN_CENTER_VERTICAL);
     sunlightSizer->AddSpacer(boxPadding);
@@ -276,6 +282,7 @@ wxWindow* RunPage::createRaymarcherBox(wxPanel* parent) {
         GuiSettings& gui = controller->getParams();
         gui.set(GuiSettingsId::SURFACE_AMBIENT, value);
         controller->tryRedraw();
+        return true;
     };
     ambientSizer->Add(ambientCtrl, 1, wxALIGN_CENTER_VERTICAL);
     ambientSizer->AddSpacer(boxPadding);
@@ -291,6 +298,7 @@ wxWindow* RunPage::createRaymarcherBox(wxPanel* parent) {
         GuiSettings& gui = controller->getParams();
         gui.set(GuiSettingsId::SURFACE_EMISSION, value);
         controller->tryRedraw();
+        return true;
     };
     emissionSizer->Add(emissionCtrl, 1, wxALIGN_CENTER_VERTICAL);
     emissionSizer->AddSpacer(boxPadding);
@@ -315,6 +323,7 @@ wxWindow* RunPage::createVolumeBox(wxPanel* parent) {
         // value in spinner is in [km^-1]
         gui.set(GuiSettingsId::VOLUME_EMISSION, value / 1.e3_f);
         controller->tryRedraw();
+        return true;
     };
     emissionSizer->Add(emissionCtrl, 1, wxALIGN_CENTER_VERTICAL);
     emissionSizer->AddSpacer(boxPadding);
@@ -331,6 +340,7 @@ wxWindow* RunPage::createVolumeBox(wxPanel* parent) {
         // value in spinner is in [km^-1]
         gui.set(GuiSettingsId::VOLUME_ABSORPTION, value / 1.e3_f);
         controller->tryRedraw();
+        return true;
     };
     absorptionSizer->Add(absorptionCtrl, 1, wxALIGN_CENTER_VERTICAL);
     absorptionSizer->AddSpacer(boxPadding);
@@ -346,6 +356,7 @@ wxWindow* RunPage::createVolumeBox(wxPanel* parent) {
         GuiSettings& gui = controller->getParams();
         gui.set(GuiSettingsId::COLORMAP_LOGARITHMIC_FACTOR, value);
         controller->tryRedraw();
+        return true;
     };
     factorSizer->Add(factorCtrl, 1, wxALIGN_CENTER_VERTICAL);
     factorSizer->AddSpacer(boxPadding);